+++ R/FullConditionals.R
@@ -797,20 +797,30 @@
   # Ledermann Bound
 #' Ledermann Bound
 #'
-#' Returns the maximum possible number of latent factors in a factor analysis model for data of dimension \code{P}. This Ledermann bound is given by the largest integer smaller than or equal to the solution \eqn{k}{k} of \eqn{(M - k)^2 \geq M + k}{(M - k)^2 >= M + k}.
+#' Returns the maximum number of latent factors in a factor analysis model for data of dimension \code{P} which actually achieves dimension reduction in terms of the number of covariance parameters. This Ledermann bound is given by the largest integer smaller than or equal to the solution \eqn{k}{k} of \eqn{(M - k)^2 \geq M + k}{(M - k)^2 >= M + k}.
 #' @param P Integer number of variables in data set. This argument is vectorized.
+#' @param isotropic Logical indicating whether uniquenesses are constrained to be isotropic, in which case the bound is simply \eqn{P-1}{P-1}. Defaults to \code{FALSE}.
 #'
 #' @return The Ledermann bound, a non-negative integer, or a vector of \code{length(P)} such bounds.
 #' @keywords utility
+#' @usage
+#' Ledermann(P,
+#'           isotropic = FALSE)
 #' @export
 #'
 #' @examples
 #' Ledermann(c(25, 50, 100))
-    Ledermann    <- function(P)  {
+    Ledermann    <- function(P, isotropic = FALSE) {
       if(!is.numeric(P)   ||
          any(P   <= 0, floor(P) != P))      stop("'P' must be a strictly positive integer", call.=FALSE)
-      R          <- P + 0.5 * (1 - sqrt(8 * P  + 1))
-        as.integer(floor(ifelse(1e-10 > abs(R - round(R)), round(R), R)))
+      if(length(isotropic) > 1  ||
+         !is.logical(isotropic))            stop("'isotropic' must be a single logical indicator", call.=FALSE)
+      if(isTRUE(isotropic))      {
+          as.integer(P - 1L)
+      } else      {
+        R        <- P + 0.5 * (1 - sqrt(8 * P  + 1))
+          as.integer(floor(ifelse(1e-10 > abs(R - round(R)), round(R), R)))
+      }
     }
 
   # Procrustes Transformation